python 2/3 compatible, localize strings, removed categories and top 50 due to site no longer active, moved images local
