+++ plugin.video.ytchannels/addon.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.ytchannels" name="Youtube Channels" version="0.2.0" provider-name="natko1412,mintsoft">
+<addon id="plugin.video.ytchannels" name="Youtube Channels" version="0.3.0" provider-name="natko1412,mintsoft,yeahme49">
 	<requires>
 		<import addon="xbmc.python" version="2.26.0"/>
-		<import addon="script.module.beautifulsoup" version="3.2.1"/>
-		<import addon="script.module.youtube" optional="false" />
+		<import addon="script.module.six" version="1.11.0"/>
+		<import addon="plugin.video.youtube"/>
 	</requires>
 	<extension point="xbmc.python.pluginsource" library="default.py">
 		<provides>video</provides>
