+++ plugin.video.ytchannels/changelog.txt
@@ -1,14 +1,9 @@
 ===============================
-Youtube Channels addon by mintsoft/natko1412
+Youtube Channels addon by natko1412
 ===============================
-
-v0.2.0 [05-July-2020]
--Patched to support user provided API Keys
--Fixed for Kodi Eden Support
--Now maintained here: https://github.com/mintsoft/kodi.plugin.ytchannels
-
 v0.1.0 [26-May-2015]
 -added support for channel playlists
+
 
 v0.0.31 [25-May-2015]
 -use unicode instead of ascii (fix arabix symbols)
@@ -27,7 +22,7 @@ v0.0.22 [20-May-2015]
 
 
 v0.0.21 [18-May-2015]
--you can now add channel by searching or entering a username
+-you can now add channel by searching or entering a username 
 -added "TOP Channels" and "Categories" (by views) charts from vidstatx.com
 -added my icons
 -fixed videos (now able to play like a playlist)
@@ -35,3 +30,7 @@ v0.0.21 [18-May-2015]
 
 v0.0.1 [17-May-2015]
 -initial release
+
+
+
+
